framework/online/non_AI_module/recommender/requirements.txt
absl-py==0.12.0
astor==0.8.1
cached-property==1.5.2
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
gast==0.2.2
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.37.1
idna==2.10
importlib-metadata==4.0.1
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==1.1.1
neo4j-connector==1.1.0
nltk==3.6.2
oauthlib==3.1.0
opt-einsum==3.3.0
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.4.1
six==1.15.0
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
tensorflow-serving-api==2.1.0
termcolor==1.1.0
tqdm==4.60.0
typing-extensions==3.10.0.0
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.1
